refactor: view 위치 리팩토링
