+++ app/src/main/java/woowacourse/omok/controller/OMokConsoleController.kt
@@ -1,7 +1,7 @@
 package woowacourse.omok.controller
 
 import woowacourse.omok.model.GameTurn
-import woowacourse.omok.model.board.LocalBoard
+import woowacourse.omok.view.LocalBoard
 import woowacourse.omok.model.player.BlackPlayer
 import woowacourse.omok.model.player.Player
 import woowacourse.omok.model.player.WhitePlayer
